requirements.txt
django>=3.1
django-cors-middleware>=1.3.1
django-debug-toolbar>=2.0
django-cors-headers>=3.5.0
djangorestframework>=3.11.0
djangorestframework-simplejwt>=4.3.0
sentry-sdk==0.14.4
pyotp>=2.3.0
gunicorn==19.9.0
psycopg2>=2.7.7
psycopg2-binary>=2.7.7
Pillow>=6.2.1
whitenoise>=3.2
honcho>=1.0.1
mysqlclient>=2.0.1
requests>=2.24.0